changed for more class inheritance: reputationwealt->threshold->{static; dinamic}->saboteur
